requirements.txt
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-grappelli==4.0.1
django-jazzmin==3.0.0
django-jet==1.0.8
django-rest-framework==0.1.0
django-suit==0.2.28
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
packaging==24.1
pillow==10.4.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python-telegram-bot==21.6
pytz==2024.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
tz==0.2.2
